Send Pull Request

+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/DuplicateIDCacheImpl.java
@@ -301,10 +301,12 @@ public class DuplicateIDCacheImpl implements DuplicateIDCache {
    public void clear() throws Exception {
       logger.debug("DuplicateIDCacheImpl(" + this.address + ")::clear removing duplicate ID data");
       synchronized (this) {
-         if (ids.size() > 0) {
+         if (ids.size() > 0 && persist) {
             long tx = storageManager.generateID();
             for (Pair<ByteArrayHolder, Long> id : ids) {
-               storageManager.deleteDuplicateIDTransactional(tx, id.getB());
+               if (id != null) {
+                  storageManager.deleteDuplicateIDTransactional(tx, id.getB());
+               }
             }
             storageManager.commit(tx);
          }
